# Added output of MEC statistics.